requirements.txt
aiogram==3.18.0
python-dotenv==1.0.1
aiosqlite==0.20.0
pytest==8.3.5
pytest-asyncio==0.25.3
